Add License, fix naming conventions, and update package configurations for NuGet publishing
