+++ DownloaderV3.sln
@@ -21,7 +21,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DownloaderV3.Source.Covalen
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DownloaderV3.Source.CovalentDocument.Tests", "test\DownloaderV3.Source.CovalentDocument.Tests\DownloaderV3.Source.CovalentDocument.Tests.csproj", "{25A75642-1115-4F2B-ADDC-BB611E16A4A9}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DownloadV3.Destination.Tests", "DownloadV3.Destination.Tests\DownloadV3.Destination.Tests.csproj", "{40922FAE-9956-446B-B798-8E2508FD8C99}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DownloaderV3.Destination.Tests", "test\DownloaderV3.Destination.Tests\DownloaderV3.Destination.Tests.csproj", "{72180384-1C77-4D91-98E4-96CBF1BB22CF}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -57,10 +57,10 @@ Global
 		{25A75642-1115-4F2B-ADDC-BB611E16A4A9}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{25A75642-1115-4F2B-ADDC-BB611E16A4A9}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{25A75642-1115-4F2B-ADDC-BB611E16A4A9}.Release|Any CPU.Build.0 = Release|Any CPU
-		{40922FAE-9956-446B-B798-8E2508FD8C99}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{40922FAE-9956-446B-B798-8E2508FD8C99}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{40922FAE-9956-446B-B798-8E2508FD8C99}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{40922FAE-9956-446B-B798-8E2508FD8C99}.Release|Any CPU.Build.0 = Release|Any CPU
+		{72180384-1C77-4D91-98E4-96CBF1BB22CF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{72180384-1C77-4D91-98E4-96CBF1BB22CF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{72180384-1C77-4D91-98E4-96CBF1BB22CF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{72180384-1C77-4D91-98E4-96CBF1BB22CF}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -73,7 +73,7 @@ Global
 		{2B72FA06-B3D5-4FBA-A44B-72BC7B5ECA7D} = {32482522-1A4F-4BCF-9418-4681E7D3A9B0}
 		{2C80420B-89C2-4486-B566-10F3E86CF2BF} = {32482522-1A4F-4BCF-9418-4681E7D3A9B0}
 		{25A75642-1115-4F2B-ADDC-BB611E16A4A9} = {800C2D6F-B4AD-4C8C-8520-8C0FCF7365B2}
-		{40922FAE-9956-446B-B798-8E2508FD8C99} = {800C2D6F-B4AD-4C8C-8520-8C0FCF7365B2}
+		{72180384-1C77-4D91-98E4-96CBF1BB22CF} = {800C2D6F-B4AD-4C8C-8520-8C0FCF7365B2}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {6596EE9D-441C-42C4-AEDA-C520A6C06544}
