+++ test/DownloaderV3.Destination.Tests/DownloaderV3.Destination.Tests.csproj
@@ -19,7 +19,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\src\DownloaderV3.Destination\DownloaderV3.Destination.csproj" />
+    <ProjectReference Include="..\..\src\DownloaderV3.Destination\DownloaderV3.Destination.csproj" />
   </ItemGroup>
 
   <ItemGroup>
